Refactor utilities library names to reduce util-ception
