+++ packages/airnode-utilities/src/arrays/index.ts
@@ -1,0 +1,1 @@
+export * from './array';
