+++ packages/airnode-utilities/src/date/date.test.ts
@@ -1,4 +1,4 @@
-import * as dateUtils from './date-utils';
+import * as dateUtils from './date';
 
 describe('formatDateTime', () => {
   it('returns the formatted string', () => {
