+++ packages/airnode-utilities/src/date/index.ts
@@ -1,0 +1,1 @@
+export * from './date';
