+++ packages/airnode-utilities/src/index.ts
@@ -1,7 +1,6 @@
+export * from './objects';
+export * from './arrays';
+export * from './date';
+export * from './promises';
+export * from './strings';
 export * from './logging';
-export * from './utils/object-utils';
-export * from './utils/array-utils';
-export * from './utils/date-utils';
-export * from './utils/promise-utils';
-export * from './utils/string-utils';
-export * from './logging-types';
