+++ packages/airnode-utilities/src/logging/index.ts
@@ -1,0 +1,2 @@
+export * from './logging';
+export * from './types';
