+++ packages/airnode-utilities/src/objects/index.ts
@@ -1,0 +1,1 @@
+export * from './object';
