+++ packages/airnode-utilities/src/objects/object.ts
