+++ packages/airnode-utilities/src/promises/index.ts
@@ -1,0 +1,1 @@
+export * from './promise';
