+++ packages/airnode-utilities/src/promises/promise.ts
