+++ packages/airnode-utilities/src/strings/string.ts
