--- packages/airnode-utilities/src/utils/index.ts
@@ -1,5 +1,0 @@
-export * from './array-utils';
-export * from './date-utils';
-export * from './object-utils';
-export * from './promise-utils';
-export * from './string-utils';
